Create files and establish base layout
Created the initial html and css files for the construction of the
Admin Dashboard project. Defined the first grid column to be 300px and
the second column to take the remaining horizontal space. Applied border
for development to keep an eye on where each html element resides, will
be removed later on in the project.