Complete layout
Completed the overall layout of the admin dashboard page using multiple
grids. Reconfigured layout to have a separate header from the main
content instead of including it within the main content parent. Set side
bar to span the height of the main container grid. Added elements to
each section to structure the rest of the page and provide parent
elements for future content.