Design projects and announcements sections
Missed a commit after design of the projects
section was completed, but both the projects grid
and the announcements section are mostly complete.
Structure is set up and is somewhat responsive
(projects grid responds well, announcements grow
to be way too tall) but still have some recoloring
and slight style adjustments to make before being
complete with the project. Next stage is to design
the trending section with a couple of fake users
then final color changes and polishing to complete
the project.